Fix science CI architecture checks
