+++ src/main/java/de/tum/cit/aet/artemis/atlas/dto/ScienceConsentUpdateDTO.java
@@ -1,4 +1,7 @@
 package de.tum.cit.aet.artemis.atlas.dto;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
+
+@JsonInclude(JsonInclude.Include.NON_EMPTY)
 public record ScienceConsentUpdateDTO(boolean active) {
 }
